Vaibhav 07/05/2024

+++ api/routes/AddDocumens/AddDocuments.js
@@ -2,6 +2,8 @@ const express = require("express");
 const router = express.Router();
 const moment = require("moment");
 const AddDocuments = require("../../models/AddDocuments/AddDocuments");
+const File_Uplode = require("../../models/File/File_Uplode");
+const Loan_Documents = require("../../models/Loan/Loan_Documents");
 
 // Post Documents
 router.post("/", async (req, res) => {
@@ -40,10 +42,10 @@ router.post("/", async (req, res) => {
   }
 });
 
-// Get Documents
+// Get Documents (// Without pagination)
 router.get("/", async (req, res) => {
   try {
-    const data = await AddDocuments.find({});
+    const data = await AddDocuments.find({}).sort({ updatedAt: -1 });
     if (data.length === 0) {
       // If no data found
       return res.status(201).json({
@@ -63,6 +65,57 @@ router.get("/", async (req, res) => {
     });
   }
 });
+
+
+// With pagination
+// router.get("/", async (req, res) => {
+//   try {
+//     // Pagination parameters
+//     const page = parseInt(req.query.page) || 1;
+//     console.log(page, "page")
+//     const limit = parseInt(req.query.limit) || 10; // Default limit to 10 documents per page
+
+//     // Calculate skip value
+//     const skip = (page - 1) * limit;
+
+//     // Fetch documents with pagination
+//     const data = await AddDocuments.find({})
+//       .sort({ updatedAt: -1 })
+//       .skip(skip)
+//       .limit(limit);
+
+//     // Count total documents
+//     const totalCount = await AddDocuments.countDocuments();
+
+//     if (data.length === 0) {
+//       // If no data found
+//       return res.status(200).json({
+//         statusCode: 200,
+//         message: "No data found",
+//       });
+//     }
+
+//     // Calculate total pages
+//     const totalPages = Math.ceil(totalCount / limit);
+
+//     res.json({
+//       success: true,
+//       data,
+//       totalPages,
+//       currentPage: page,
+//       totalCount,
+//     });
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({
+//       success: false,
+//       message: "Internal Server Error",
+//     });
+//   }
+// });
+
+
+
 
 // Update Documents
 router.put("/:document_id", async (req, res) => {
@@ -101,6 +154,21 @@ router.put("/:document_id", async (req, res) => {
 router.delete("/:document_id", async (req, res) => {
   try {
     const { document_id } = req.params;
+
+    const documentExistsInFileUploads = await File_Uplode.findOne({
+      "documents.loan_document_id": document_id,
+    });
+
+    const documentExistsInLoanDocuments = await Loan_Documents.findOne({
+      document_ids: document_id,
+    });
+
+    if (documentExistsInFileUploads || documentExistsInLoanDocuments) {
+      return res.status(200).json({
+        statusCode: 201,
+        message: "Document cannot be deleted because it is currently in use.",
+      });
+    }
 
     const deletedDocument = await AddDocuments.findOneAndDelete({
       document_id: document_id,
